convert remaining unittest-style tests to pytest
