+++ tests/integration/renderer/test_race_conditions.py
@@ -21,13 +21,13 @@ routes = [
 ]
 
 permuted_strs = [
-    ','.join(endpoints) for endpoints in itertools.permutations(routes, len(routes))
+    ",".join(endpoints) for endpoints in itertools.permutations(routes, len(routes))
 ]
 
 
 @pytest.mark.parametrize("permuted_str", permuted_strs)
 def test_rdrc001_race_conditions(dash_duo, permuted_str):
-    endpoints = permuted_str.split(',')
+    endpoints = permuted_str.split(",")
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -49,9 +49,7 @@ def test_rdrc001_race_conditions(dash_duo, permuted_str):
     dash_duo.start_server(app)
 
     dash_duo.wait_for_text_to_equal(
-        '#output',
-        "initial value",
-        timeout=DELAY_TIME * (len(endpoints) + 3) + 3
+        "#output", "initial value", timeout=DELAY_TIME * (len(endpoints) + 3) + 3
     )
 
     assert not dash_duo.get_logs()
